디플로이시 리눅스에서 파일 불러오지 못하는 에러 수정
The problem was incorrect case for components. It works on Windows and
macOS (that’s why it build successfully on my Windows laptop and you
might be using either of the two), but, Netlify uses Ubuntu which is
Linux based and Linux can’t work with the mixed cases. So, all your
components are in the folders which start with a small case, while the
pages start with a capital case. While importing the components, you
need to maintain the case, but, instead, you were importing all the
components with the starting letter capital.

https://answers.netlify.com/t/build-command-failed-error-refers-to-webpack/25012/6